requirements.txt
# Requirements for Any App

# AI Providers
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
cohere>=5.0.0
groq>=0.4.0
mistralai>=0.1.0

# GUI
customtkinter>=5.2.0
Pillow>=10.0.0

# Vision (captura de pantalla)
pyautogui>=0.9.54

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
